[RMAX_APPS_LIB] Align with Rivermax v1.30.16
Rivermax release date: April 2023

Rivermax application library is a high-level C++ library for the
Rivermax library. It wraps the Rivermax low-level API and provides ease
of use use and services for an application developer. The user can
extend the library based on need.

The Rivermax application library is supported for both Windows and Linux
with all Rivermax official distributions.

Current applications:
- rmax_ipo_receiver
- rmax_xstream_generic_sender
- rmax_xstream_media_sender

Signed-off-by: Ilan Smith <ilansm@nvidia.com>