[APPS] Handle the race between IPMX SRs and first frame packets
RM4414971

A Sender Report (SR) for frame N+1 should be sent somewhere
in the middle of frame N. But due to delay in polling,
this SR can be received before beginning of the frame N.
When frame N first packet is received we use a previous version of SR,
matching this frame. In any case (with good or bad order)
the sender report with the matching RTP timestamp is used.

Signed-off-by: Sergei Maksimenko <smaksimenko@nvidia.com>