[APPS] Add Media Probe application (video and alpha/key receiver)
RM4579253

This is an example of application that uses Rivermax API to receive the
Main Essence (color) and Alpha/Key channels of a video signal.
These two components are received as separate ST2110-20 RTP streams.
The application tracks the synchronization of these components by matching
RTP timestamps across the two streams.
The application also tracks:
 - the media delay of each component by calculating
   the difference between the receive timestamp and the RTP timestamp.
 - the number of dropped packets by calculating tracking RTP sequence numbers.
 - the number of received frames, packets, bytes.
The application is capable of receiving multiple video and alpha/key streams.

Signed-off-by: Sergei Maksimenko <smaksimenko@nvidia.com>